pyproject.toml
[tool.poetry]
name = "kursovaya-5"
version = "0.1.0"
description = ""
authors = ["Vadim Kravcov <v.sem4ik@mail.ru>"]
readme = "README.md"
packages = [{include = "kursovaya_5"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
psycopg2-binary = "^2.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
